+++ edu.nd.dronology.core/src/edu/nd/dronology/core/vehicle/commands/TakeoffCommand.java
@@ -2,7 +2,7 @@ package edu.nd.dronology.core.vehicle.commands;
 
 public class TakeoffCommand extends AbstractDroneCommand implements IDroneCommand {
 
-	public TakeoffCommand(String droneId, long altitude) {
+	public TakeoffCommand(String droneId, double altitude) {
 		super(droneId, CommandIds.TAKEOFF_COMMAND);
 		innerdata.put("altitude", altitude);
 	}
